fix warnings from old Intel compiler
git-svn-id: https://qthreads.googlecode.com/svn/trunk@2093 9dbd018f-d1da-d342-4c11-58c73dfe904b

+++ src/io.c
@@ -114,7 +114,7 @@ void INTERNAL qt_process_blocking_calls(void)
     /* do something with <item> */
     switch(item->op) {
         default:
-            fprintf(stderr, "Unhandled syscall: %i\n", item->op);
+            fprintf(stderr, "Unhandled syscall: %u\n", (unsigned int)item->op);
             abort();
 #if HAVE_DECL_SYS_ACCEPT
         case ACCEPT:
